+++ http/src/main/scala/com/twitter/finatra/http/filters/StatsFilter.scala
@@ -11,7 +11,8 @@ import com.twitter.util._
 import javax.inject.{Inject, Singleton}
 
 private object StatsFilter {
-  val url = "https://twitter.github.io/finatra/user-guide/http/exceptions.html#exceptionmappingfilter"
+  val url =
+    "https://twitter.github.io/finatra/user-guide/http/exceptions.html#exceptionmappingfilter"
 
   /** INTENDED FOR INTERNAL USE ONLY */
   object Stats {
@@ -25,7 +26,8 @@ private object StatsFilter {
         statusCodeTime = statsReceiver.scope("time").stat(statusCode.toString),
         statusClassTime = statsReceiver.scope("time").stat(statusClass),
         successCount = if (perEndpoint) Some(statsReceiver.counter("success")) else None,
-        failuresCount = if (perEndpoint) Some(statsReceiver.counter("failures")) else None
+        failuresCount = if (perEndpoint) Some(statsReceiver.counter("failures")) else None,
+        ignoredCount = if (perEndpoint) Some(statsReceiver.counter("ignored")) else None
       )
     }
   }
@@ -39,9 +41,9 @@ private object StatsFilter {
     statusCodeTime: Stat,
     statusClassTime: Stat,
     successCount: Option[Counter],
-    failuresCount: Option[Counter]
-  ) {
-    def count(duration: Duration, response: Response, success: Boolean): Unit = {
+    failuresCount: Option[Counter],
+    ignoredCount: Option[Counter]) {
+    def count(duration: Duration, response: Response, responseClass: ResponseClass): Unit = {
       requestCount.foreach(_.incr())
       statusCodeCount.incr()
       statusClassCount.incr()
@@ -51,10 +53,10 @@ private object StatsFilter {
       statusCodeTime.add(durationMs.toFloat)
       statusClassTime.add(durationMs.toFloat)
 
-      if (success) {
-        successCount.foreach(_.incr())
-      } else {
-        failuresCount.foreach(_.incr())
+      responseClass match {
+        case ResponseClass.Ignorable => ignoredCount.foreach(_.incr())
+        case ResponseClass.Failed(_) => failuresCount.foreach(_.incr())
+        case ResponseClass.Successful(_) => successCount.foreach(_.incr())
       }
     }
   }
@@ -85,6 +87,7 @@ private object StatsFilter {
  *
  * {{{
  *   route/foo/GET/failures 1
+ *   route/foo/GET/ignored 0
  *   route/foo/GET/requests 1
  *   route/foo/GET/status/500 1
  *   route/foo/GET/status/5XX 1
@@ -98,13 +101,33 @@ private object StatsFilter {
  *   time/5XX 86.000000 [86.0]
  * }}}
  *
+ * Example stats for a failed GET request to a route named `/foo` which classifies
+ * the failure as Ignorable::
+ *
+ * {{{
+ *   route/foo/GET/failures 0
+ *   route/foo/GET/ignored 1
+ *   route/foo/GET/requests 1
+ *   route/foo/GET/status/500 0
+ *   route/foo/GET/status/5XX 0
+ *   route/foo/GET/success 0
+ *   route/foo/GET/time 86.000000 [86.0]
+ *   route/foo/GET/time/500 86.000000 [86.0]
+ *   route/foo/GET/time/5XX 86.000000 [86.0]
+ *   status/500 0
+ *   status/5XX 0
+ *   time/500 86.000000 [86.0]
+ *   time/5XX 86.000000 [86.0]
+ * }}}
+ *
+ * requests == success + failures + ignored
+ * The logical success rate for a method can be calculated as `success / (success + failures)`
+ *
  * @note It is expected that this Filter occurs "BEFORE" the [[ExceptionMappingFilter]] in a
  *       given filter chain, e.g., `StatsFilter.andThen(ExceptionMappingFilter)`. It is expected
  *       that there SHOULD be a returned response because the [[ExceptionMappingFilter]] should
  *       return give a non-exception response.
- *
  * @see [[com.twitter.finagle.http.filter.StatsFilter Finagle HTTP StatsFilter]]
- *
  * @param statsReceiver      the [[com.twitter.finagle.stats.StatsReceiver]] to which
  *                           to record stats.
  * @param responseClassifier an [[HttpResponseClassifier]] used to determine when a response
@@ -113,11 +136,10 @@ private object StatsFilter {
  *           [[com.twitter.finagle.http.Request]] type.
  */
 @Singleton
-class StatsFilter[R <: Request] @Inject()(
-  statsReceiver: StatsReceiver,
-  responseClassifier: HttpResponseClassifier
-) extends SimpleFilter[R, Response]
-  with Logging {
+class StatsFilter[
+  R <: Request] @Inject()(statsReceiver: StatsReceiver, responseClassifier: HttpResponseClassifier)
+    extends SimpleFilter[R, Response]
+    with Logging {
 
   import StatsFilter._
 
@@ -157,7 +179,8 @@ class StatsFilter[R <: Request] @Inject()(
    * is classified as a "success", we incr the exception counter(s) (in addition to the "success"
    * or "failures" counters). Conversely, if a response (non-exception) is returned which is classified
    * as a "failure", we incr the "failures" counter but we do not incr any exception counter. Finally,
-   * responses or exceptions classified as "ignorable" only increment the exception counter(s).
+   * for responses or exceptions classified as "ignorable", we incr the "ignored" counter and
+   * the exception counter(s).
    *
    * {{{
    *                   *-----------------*---------------------------*
@@ -169,7 +192,7 @@ class StatsFilter[R <: Request] @Inject()(
    *  *----------------*-----------------*---------------------------*
    *  |    FAILED      | failed.incr()   | failed.incr(), exc.incr() |
    *  *----------------|-----------------|---------------------------|
-   *  |   IGNORABLE    | (no-op)         | exc.incr()                |
+   *  |   IGNORABLE    | (no-op)         | ignored.incr(), exc.incr()|
    *  *----------------*-----------------*---------------------------*
    * }}}
    *
@@ -179,16 +202,9 @@ class StatsFilter[R <: Request] @Inject()(
   def apply(request: R, service: Service[R, Response]): Future[Response] = {
     val elapsed = Stopwatch.start()
     service(request).respond { response =>
-      val success = responseClassifier.applyOrElse(
-        ReqRep(request, response),
-        ResponseClassifier.Default
-      ) match {
-        case ResponseClass.Ignorable => // Do nothing.
-        case ResponseClass.Failed(_) =>
-          count(elapsed(), request, response, success = false)
-        case ResponseClass.Successful(_) =>
-          count(elapsed(), request, response, success = true)
-      }
+      val responseClass =
+        responseClassifier.applyOrElse(ReqRep(request, response), ResponseClassifier.Default)
+      count(elapsed(), request, response, responseClass = responseClass)
     }
   }
 
@@ -198,7 +214,7 @@ class StatsFilter[R <: Request] @Inject()(
     duration: Duration,
     request: Request,
     tryResponse: Try[Response],
-    success: Boolean
+    responseClass: ResponseClass
   ): Unit = {
     val response = tryResponse match {
       case Return(v) => v
@@ -211,10 +227,17 @@ class StatsFilter[R <: Request] @Inject()(
         // Treat exceptions as empty 500 errors
         SimpleResponse(Status.InternalServerError)
     }
-    globalStats(response.statusCode).count(duration, response, success)
+    // Only increment global stats if the response is Successful or Failed, not Ignorable.
+    responseClass match {
+      case ResponseClass.Successful(_) =>
+        globalStats(response.statusCode).count(duration, response, responseClass)
+      case ResponseClass.Failed(_) =>
+        globalStats(response.statusCode).count(duration, response, responseClass)
+      case _ => // Do Nothing
+    }
     RouteInfo(request).foreach { routeInfo =>
       perRouteStats((routeInfo, request.method, response.statusCode))
-        .count(duration, response, success)
+        .count(duration, response, responseClass)
     }
   }
 }