+++ http/src/test/scala/com/twitter/finatra/http/tests/filters/StatsFilterTest.scala
@@ -153,6 +153,48 @@ class StatsFilterTest extends Test with Mockito {
     }
   }
 
+  test("per route failed request classified as ignorable") {
+    val httpResponseClassifier = HttpResponseClassifier(
+      ResponseClassifier.named("TestClassifier") {
+        case ReqRep(_, Throw(t)) if t.isInstanceOf[IllegalArgumentException] =>
+          ResponseClass.Ignored
+      }
+    )
+
+    val statsFilter = new StatsFilter[Request](
+      statsReceiver,
+      httpResponseClassifier
+    )
+
+    val request = Request()
+    val service = Service.mk[Request, Response] { _ =>
+      Future.exception(new IllegalArgumentException("oops"))
+    }
+
+    RouteInfo.set(request, routeInfo)
+    withService(service) {
+      Await.ready(statsFilter.apply(request, service), 2.seconds)
+      /* only global counters and histos */
+      assert(statsReceiver.counters.get(Seq("status", "5XX")).isEmpty)
+      assert(statsReceiver.counters.get(Seq("status", "2XX")).isEmpty)
+      assert(statsReceiver.stats.get(Seq("time", "5XX")).isEmpty)
+      assert(statsReceiver.stats.get(Seq("time", "2XX")).isEmpty)
+
+      /* per-route counters */
+      statsReceiver.counters(List("route", "foo", "GET", "requests")) should equal(1)
+      statsReceiver.counters(List("route", "foo", "GET", "status", "5XX")) should equal(1)
+      statsReceiver.counters(List("route", "foo", "GET", "status", "500")) should equal(1)
+      statsReceiver.counters(List("route", "foo", "GET", "ignored")) should equal(1) // server errors ignored
+      statsReceiver.counters(List("route", "foo", "GET", "failures")) should equal(0)
+      statsReceiver.counters(List("route", "foo", "GET", "success")) should equal(0)
+
+      /* per-route histos */
+      statsReceiver.stats.get(List("route", "foo", "GET", "time")) should not be None
+      statsReceiver.stats.get(List("route", "foo", "GET", "time", "5XX")) should not be None
+      statsReceiver.stats.get(List("route", "foo", "GET", "time", "500")) should not be None
+    }
+  }
+
   test("per route IllegalArgumentException classified as success") {
     val httpResponseClassifier = HttpResponseClassifier(
       ResponseClassifier.named("TestClassifier") {
@@ -179,6 +221,7 @@ class StatsFilterTest extends Test with Mockito {
       statsReceiver.counters(List("route", "foo", "GET", "requests")) should equal(1)
       statsReceiver.counters(List("route", "foo", "GET", "status", "5XX")) should equal(1)
       statsReceiver.counters(List("route", "foo", "GET", "status", "500")) should equal(1)
+      statsReceiver.counters(List("route", "foo", "GET", "ignored")) should equal(0) // IllegalArgumentException classified as a success
       statsReceiver.counters(List("route", "foo", "GET", "failures")) should equal(0) // IllegalArgumentException classified as a success
       statsReceiver.counters(List("route", "foo", "GET", "success")) should equal(1)
     }
@@ -210,6 +253,7 @@ class StatsFilterTest extends Test with Mockito {
       statsReceiver.counters(List("route", "foo", "GET", "requests")) should equal(1)
       statsReceiver.counters(List("route", "foo", "GET", "status", "2XX")) should equal(1)
       statsReceiver.counters(List("route", "foo", "GET", "status", "200")) should equal(1)
+      statsReceiver.counters(List("route", "foo", "GET", "ignored")) should equal(0)
       statsReceiver.counters(List("route", "foo", "GET", "failures")) should equal(1) // 200 OK response but classified as a "failure"
       statsReceiver.counters(List("route", "foo", "GET", "success")) should equal(0)
     }
@@ -241,6 +285,7 @@ class StatsFilterTest extends Test with Mockito {
       statsReceiver.counters(List("route", "foo", "GET", "requests")) should equal(1)
       statsReceiver.counters(List("route", "foo", "GET", "status", "5XX")) should equal(1)
       statsReceiver.counters(List("route", "foo", "GET", "status", "500")) should equal(1)
+      statsReceiver.counters(List("route", "foo", "GET", "ignored")) should equal(0) // Any 5xx is a success
       statsReceiver.counters(List("route", "foo", "GET", "failures")) should equal(0) // Any 5xx is a success
       statsReceiver.counters(List("route", "foo", "GET", "success")) should equal(1)
     }